organize ux-capture src/ and examples/, isolate babel config per package
