+++ package.json
@@ -8,6 +8,7 @@
   "license": "MIT",
   "private": true,
   "scripts": {
+    "build": "lerna run build",
     "do-version": "lerna version",
     "do-publish": "lerna publish from-package --yes"
   },
@@ -20,10 +21,11 @@
     "@babel/cli": "^7.2.3",
     "@babel/core": "^7.2.2",
     "@babel/plugin-proposal-class-properties": "^7.2.3",
+    "@babel/plugin-transform-flow-strip-types": "^7.2.3",
     "@babel/preset-env": "^7.2.3",
-    "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "babel-eslint": "9.0.0",
+    "babel-loader": "^8.0.4",
     "eslint": "5.6.0",
     "eslint-config-prettier": "^3.3.0",
     "eslint-loader": "^2.1.1",
@@ -39,7 +41,8 @@
     "semver": "^5.6.0",
     "uglify-es": "^3.3.9",
     "usertiming": "^0.1.8",
-    "webpack": "4.19.1"
+    "webpack": "4.19.1",
+    "webpack-cli": "^3.1.2"
   },
   "husky": {
     "hooks": {
