+++ packages/react-ux-capture/.babelrc.js
@@ -2,7 +2,9 @@ module.exports = {
 	presets: [
 		'@babel/preset-react',
 		['@babel/preset-env', { targets: { browsers: ['last 2 versions'] } }],
-		'@babel/preset-flow',
 	],
-	plugins: ['@babel/plugin-proposal-class-properties'],
+	plugins: [
+		'@babel/plugin-transform-flow-strip-types',
+		'@babel/plugin-proposal-class-properties',
+	],
 };
