+++ packages/ux-capture/examples/kitten.jpg
