+++ packages/ux-capture/src/View.js
