+++ packages/ux-capture/src/ux-capture.js
