+++ packages/ux-capture/webpack.config.js
